rename controlled dropdown and extract controlled multiple select input
